requirements.txt
streamlit>=1.32.0
tensorflow-cpu>=2.16.1
mysql-connector-python
Pillow
gdown
requests
numpy
protobuf!=4.24.0,>=3.19.6